[[ REFACTOR ]] Refactor code on input_data
